fix: template-react-ts content script hmr not working because of bad file extension
