+++ template-react-ts/src/content/main.tsx
@@ -1,0 +1,14 @@
+import { StrictMode } from 'react'
+import { createRoot } from 'react-dom/client'
+import App from './views/App.js'
+
+console.log('[CRXJS] Hello world from content script!')
+
+const container = document.createElement('div')
+container.id = 'crxjs-app'
+document.body.appendChild(container)
+createRoot(container).render(
+  <StrictMode>
+    <App />
+  </StrictMode>,
+)
